ISECL-14458: SKC Binary deployment scripts fix

+++ build/skc-tools/sgx_agent/build_scripts/install_sgxsdk.sh
@@ -19,6 +19,10 @@ install_sgxsdk()
 	chmod u+x sgx_linux_x64_sdk_$SGX_SDK_VERSION.bin
 	./sgx_linux_x64_sdk*.bin -prefix=$SGX_INSTALL_DIR || exit 1
 	source $SGX_INSTALL_DIR/sgxsdk/environment
+	if [ $? -ne 0 ]; then
+		echo "failed while setting sgx environment"
+		exit 1
+	fi
 	rm -f sgx_linux_x64_sdk_$SGX_SDK_VERSION.bin
 }
