+++ build/skc-tools/skc_library/build_scripts/install_sgxsdk.sh
@@ -21,6 +21,10 @@ install_sgxsdk()
 	chmod +x sgx_linux_x64_sdk_$SGX_SDK_VERSION.bin
 	./sgx_linux_x64_sdk_$SGX_SDK_VERSION.bin -prefix=$SGX_INSTALL_DIR || exit 1
 	source $SGX_INSTALL_DIR/sgxsdk/environment
+	if [ $? -ne 0 ]; then
+		echo "failed while setting sgx environment"
+		exit 1
+	fi
 	rm -rf sgx_linux_x64_sdk_$SGX_SDK_VERSION.bin
 }
